기존 예시코드 정리
