+++ renderer/interfaces/index.ts
@@ -14,8 +14,3 @@ declare global {
     }
   }
 }
-
-export type User = {
-  id: number;
-  name: string;
-};
