--- renderer/pages/about.tsx
@@ -1,14 +1,0 @@
-import Link from 'next/link';
-import Layout from '../components/Layout';
-
-const AboutPage = () => (
-  <Layout title="About | Next.js + TypeScript + Electron Example">
-    <h1>About</h1>
-    <p>This is the about page</p>
-    <p>
-      <Link href="/">Go home</Link>
-    </p>
-  </Layout>
-);
-
-export default AboutPage;
