--- renderer/utils/sample-api.ts
@@ -1,34 +1,0 @@
-import { User } from '../interfaces/index';
-
-/** Dummy user data. */
-export const dataArray: User[] = [
-  { id: 101, name: 'Alice' },
-  { id: 102, name: 'Bob' },
-  { id: 103, name: 'Caroline' },
-  { id: 104, name: 'Dave' },
-];
-
-/**
- * Calls a mock API which finds a user by ID from the list above.
- *
- * Throws an error if not found.
- */
-export async function findData(id: number | string) {
-  const selected = dataArray.find((data) => data.id === Number(id));
-
-  if (!selected) {
-    throw new Error('Cannot find user');
-  }
-
-  return selected;
-}
-
-/** Calls a mock API which returns the above array to simulate "get all". */
-export async function findAll() {
-  // Throw an error, just for example.
-  if (!Array.isArray(dataArray)) {
-    throw new Error('Cannot find users');
-  }
-
-  return dataArray;
-}
